# backend/public/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend Orders Rest API Description",
    "version": "1.0.0",
    "title": "Backend Order APIs Docs",
    "contact": {
      "email": "ketan.mittal@nagarro.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "",
  "basePath": "",
  "tags": [
    {
      "name": "orders",
      "description": "Orders Rest API"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {    
    "/orders": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Create a new order",
        "description": "Create a new order",
        "operationId": "createNewOrder",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Order object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Orders"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid input request"
          }
        }
      },
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get orders list",
        "description": "Get orders list",
        "operationId": "listorder",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "Current Page No.",
            "required": true,
            "default":1            
          },
          {
            "in": "query",
            "name": "limit",
            "description": "No. of orders to be fetched",
            "required": true,
            "default":5
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid input request"
          }
        }
      }
    },
    "/orders/{orderId}": {
      "patch": {
        "tags": [
          "orders"
        ],
        "summary": "Update an order status",
        "description": "Take order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Valid order id"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Order object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid request parameters"
          },
          "409": {
            "description": "Already taken"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Orders": {
      "type": "object",
      "required": [
        "origin",
        "destination"
      ],
      "properties": {
        "origin": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ObjectSchemaDefinition"
          }
        },
        "destination": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ObjectSchemaDefinition"
          }
        }
      }
    },
    "OrderUpdate": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "description": "Order status",
          "enum": [
            "TAKEN",
            "UNASSIGNED"
          ]
        }
      }
    },
    "ObjectSchemaDefinition": {
      "type": "string"
    }
  }
}